backend/requirements.txt
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
marshmallow==3.21.3
django-rest-marshmallow==4.0.2
pyjwt==2.6.0
kafka==1.3.5
kafka-python-ng==2.2.2
psycopg2-binary==2.9.9
Werkzeug==2.2.2
SQLAlchemy==2.0.31
arrow==1.3.0